Clean move of functions from node to maps (BROKEN CODE)
